Suppressing circular reference fk error, and introducing test to show this is currently 'broken'
It works without this but: ultimately, the foreign keys are not needed, they are there solely for SQL Alchemy to know what columns to use to auto create relationships. Data integrity does not strictly speaking need to be enforced at the database level unless another application views it directly. Currently there is no such application. The DB is here as a development db.

In any case, this isn't a problem with the framework, it is how it can be used and I am choosing to use it. Circular dependance in a DB is simply bad, so I don't overly care I'm suppressing this warning.